Changed DB to SQLlite. Updated CI GitHub Actions Workflows. Testing SecurityController

+++ src/Controller/TaskController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Task;
 use App\Form\TaskType;
 use App\Repository\TaskRepository;
+use App\Security\Voter\TaskVoter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -12,7 +13,7 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class TaskController extends AbstractController
 {
-    #[Route('/task', name: 'task_list')]
+    #[Route('/tasks', name: 'task_list')]
     public function listAction(TaskRepository $taskRepository): Response
     {
         return $this->render(
@@ -21,7 +22,7 @@ class TaskController extends AbstractController
         );
     }
 
-    #[Route('/task/create', name: 'task_create')]
+    #[Route('/tasks/create', name: 'task_create')]
     public function createAction(
         Request $request,
         TaskRepository $taskRepository
@@ -32,10 +33,12 @@ class TaskController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            // add user to save
+            $user = $this->getUser();
+            $task->setUser($user);
+
             $taskRepository->save($task, true);
 
-            $this->addFlash('success', 'La tâche a été bien été ajoutée.');
+            $this->addFlash('success', 'Task created successfully.');
 
             return $this->redirectToRoute('task_list');
         }
@@ -49,6 +52,12 @@ class TaskController extends AbstractController
         Request $request,
         TaskRepository $taskRepository
     ) {
+        $this->denyAccessUnlessGranted(
+            attribute: TaskVoter::EDIT,
+            subject: $task,
+            message: 'You shall not pass!'
+        );
+
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
@@ -56,7 +65,7 @@ class TaskController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             $taskRepository->save($task, true);
 
-            $this->addFlash('success', 'La tâche a bien été modifiée.');
+            $this->addFlash('success', 'The task has been modified.');
 
             return $this->redirectToRoute('task_list');
         }
@@ -75,7 +84,7 @@ class TaskController extends AbstractController
         $task->toggle(!$task->isDone());
         $taskRepository->save($task, true);
 
-        $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+        $this->addFlash('success', sprintf('The task %s has been modified.', $task->getTitle()));
 
         return $this->redirectToRoute('task_list');
     }
@@ -85,9 +94,15 @@ class TaskController extends AbstractController
         Task $task,
         TaskRepository $taskRepository
     ) {
+        $this->denyAccessUnlessGranted(
+            attribute: TaskVoter::DELETE,
+            subject: $task,
+            message: 'You shall not pass!'
+        );
+
         $taskRepository->remove($task, true);
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+        $this->addFlash('success', 'The task has been successfully deleted.');
 
         return $this->redirectToRoute('task_list');
     }
